+++ CHANGELOG.md
@@ -53,6 +53,8 @@
 
 - [#7096](https://github.com/ChainSafe/forest/issues/7096): `eth_subscribe` `logs` now re-emits the logs of reorg-reverted tipsets with `removed: true`, ahead of the logs of the replacing tipsets.
 
+- [#7156](https://github.com/ChainSafe/forest/pull/7156): Fixed the block-level bloom is now computed from the block's logs.
+
 ## Forest v0.33.7 "Shimmergloom"
 
 ### Added